+++ packer-images/windows-2019/scripts/Install-SQL-Server.ps1
@@ -1,1 +1,0 @@
-choco install sql-server-2019 -y